+++ common/src/main/java/immersive_furniture/MaterialTags.java
@@ -1,0 +1,15 @@
+package immersive_furniture;
+
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.level.block.Block;
+
+public final class MaterialTags {
+    public static final TagKey<Block> WOOL = create("wool");
+
+    private static TagKey<Block> create(String name) {
+        return TagKey.create(Registries.BLOCK, Common.locate(name));
+    }
+}
+