+++ common/src/main/java/immersive_furniture/client/ClientHandlerImpl.java
@@ -1,12 +1,12 @@
 package immersive_furniture.client;
 
 import immersive_furniture.ClientHandler;
-import immersive_furniture.client.gui.ArtisansWorkstationScreen;
+import immersive_furniture.client.gui.ArtisansWorkstationEditorScreen;
 import net.minecraft.client.Minecraft;
 
 public class ClientHandlerImpl implements ClientHandler {
     @Override
     public void openScreen() {
-        Minecraft.getInstance().setScreen(new ArtisansWorkstationScreen());
+        Minecraft.getInstance().setScreen(new ArtisansWorkstationEditorScreen());
     }
 }