+++ common/src/main/java/immersive_furniture/client/gui/ArtisansWorkstationLibraryScreen.java
@@ -1,0 +1,5 @@
+package immersive_furniture.client.gui;
+
+public class ArtisansWorkstationLibraryScreen extends ArtisansWorkstationScreen {
+
+}