+++ common/src/main/java/immersive_furniture/client/renderer/FurnitureBlockEntityRenderer.java
@@ -4,7 +4,7 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import immersive_furniture.block.FurnitureBlockEntity;
 import immersive_furniture.client.model.DynamicAtlas;
-import immersive_furniture.client.model.FurnitureBakedModel;
+import immersive_furniture.client.model.FurnitureModelBaker;
 import immersive_furniture.data.FurnitureData;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.MultiBufferSource;
@@ -28,7 +28,7 @@ public class FurnitureBlockEntityRenderer<T extends FurnitureBlockEntity> implem
 
         BlockRenderDispatcher blockRenderer = Minecraft.getInstance().getBlockRenderer();
         BlockState state = blockEntity.getBlockState();
-        BakedModel bakedModel = FurnitureBakedModel.getModel();
+        BakedModel bakedModel = FurnitureModelBaker.getModel(data);
         // VertexConsumer consumer = buffer.getBuffer(translucent ? Sheets.translucentCullBlockSheet() : Sheets.cutoutBlockSheet());
         ResourceLocation location = DynamicAtlas.ENTITY.getLocation();
         VertexConsumer consumer = buffer.getBuffer(translucent ? RenderType.entityTranslucent(location) : RenderType.entityCutout(location));